delete request fix

+++ CollabDo.Application/Services/EmployeeService.cs
@@ -1,5 +1,6 @@
 ﻿using CollabDo.Application.Dtos;
 using CollabDo.Application.Entities;
+using CollabDo.Application.Exceptions;
 using CollabDo.Application.IRepositories;
 using CollabDo.Application.IServices;
 using CollabDo.Application.Validation;
@@ -82,22 +83,31 @@ namespace CollabDo.Application.Services
             EmployeeValidation employeeValidation = new EmployeeValidation(_employeeRepository);
             employeeValidation.ValidateEmployeeId(employeeId);
 
-            Guid leaderId = await _userRepository.GetUserIdByEmail(leaderEmail);
+            Guid leaderUserId = await _userRepository.GetUserIdByEmail(leaderEmail);
+
+            Guid leaderId = _leaderRepository.GetLeaderId(leaderUserId);
 
             KeycloakUserRequestModel employeeUser = await _userRepository.GetUser(userId);
 
-            EmployeeRequestEntity employeeRequestEntity = _employeeRequestRepository.GetEmployeeRequest(employeeUser.Email, leaderId);
-
-            _employeeRequestRepository.DeleteEmployeeRequest(employeeRequestEntity);
-
             EmployeeEntity employeeEntity = _employeeRepository.GetEmployee(employeeId);
-            employeeEntity.LeaderRequestEmail = null;
-            employeeEntity.LeaderId = Guid.Empty;
+            employeeEntity.LeaderRequestEmail = "";
+            employeeEntity.LeaderId = null;
             employeeEntity.Leader = null;
             employeeEntity.ModifiedBy = userId;
             employeeEntity.ModifiedOn = DateTime.UtcNow;
 
             _employeeRepository.UpdateEmployee(employeeEntity);
+
+            EmployeeRequestEntity employeeRequestEntity = _employeeRequestRepository.GetEmployeeRequest(employeeUser.Email, leaderId);
+
+            if(employeeRequestEntity == null) 
+            {
+                throw new EntityNotFoundException("Request not found");
+            }
+
+            _employeeRequestRepository.DeleteEmployeeRequest(employeeRequestEntity);
+
+            
 
             return employeeEntity.Id;
         }
